Modified ui testing script
