--- community_images/ghost/bitnami/certs/server.crt
@@ -1,33 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIFwjCCA6qgAwIBAgITEuZy8dMVpTKArSFYV9dykiyK0jANBgkqhkiG9w0BAQsF
-ADBxMQswCQYDVQQGEwJVUzELMAkGA1UECAwCQ0ExEjAQBgNVBAcMCVN1bm55dmFs
-ZTESMBAGA1UECgwJUmFwaWRGb3J0MRkwFwYDVQQLDBBDb21tdW5pdHkgSW1hZ2Vz
-MRIwEAYDVQQDDAlsb2NhbGhvc3QwHhcNMjMwNTE3MTM0OTU0WhcNMzMwNTE0MTM0
-OTU0WjBxMQswCQYDVQQGEwJVUzELMAkGA1UECAwCQ0ExEjAQBgNVBAcMCVN1bm55
-dmFsZTESMBAGA1UECgwJUmFwaWRGb3J0MRkwFwYDVQQLDBBDb21tdW5pdHkgSW1h
-Z2VzMRIwEAYDVQQDDAlsb2NhbGhvc3QwggIiMA0GCSqGSIb3DQEBAQUAA4ICDwAw
-ggIKAoICAQDBmEHK9nQGeRZT5uVC9nvpbc9zui6+gfcc4LysCkPJUIYWrcbUGRbE
-2TtyEKz6xGwqFfEI9QfbCuAwnDNZ5PamT/GfP4WntUWCR5Y4v8VEwPPb09tqzuQa
-/MkmzpG84NjK5CJL31a2iO1851q3rcGJdil7xEm+hKVmBLXFKVqzWaXNirc0dbbQ
-GaecOIU1B1iFW8rSsMaaX2LMEN3yfbxVzndAKWUwc4NdA+MTzHbva8lDNy+OQCHZ
-z1z/Nd86IIyWH36/GmY1MN6UEKAsK36JtK3RMhkE6xxQv/hp4amjIc8pYYmxX0T8
-YVF6dFM8ct1cKbUv8X/dJHrIx+Xg/LoQLspkqGQE+VMjfV14Yafby9TewhIN9oek
-qGYvzmbkPodX3M3TEWyCUmiBiYY3QSax6NdpHk2mLOI1ezJfnby3rytgeDJxNXuQ
-gwc92bKJtc4IUXqQKO1DWO1rf6H85GBa2zN6WZ8RIB3mXtuDQOXRFIDN7vmh2Nmo
-6AYo9s6zegycbnO10UhdmL1zZcaNA3vCIKRshRgMjhuh6hyxv2aIf2eHn6OqCYKN
-6rqZuPB+7yuuaVJwrPLbAum6BXv3GyV0RNnPnTdqx8hBMq6NpAU/SuHp/7m5T1Pr
-cnuQe+5C6HE+i1BSxn54l2GAC8u1shefSp+gJVyDUIYkPJ4iTUa+sQIDAQABo1Mw
-UTAdBgNVHQ4EFgQUzzpzFWl9a0ENnUJOqP8MLKQpBfQwHwYDVR0jBBgwFoAUzzpz
-FWl9a0ENnUJOqP8MLKQpBfQwDwYDVR0TAQH/BAUwAwEB/zANBgkqhkiG9w0BAQsF
-AAOCAgEAGttSuk5BaohWNWwds+iaH73a6qmdHTm6XZJC7dmIAVI4pZIWg0Cu7uiq
-uPS0r8U+hKmsUWuXeVlWqeaP5X8eF4EzoD8F0nvEDUhpAIc8jRsrDgVoA0e0hFNy
-IJ49BNMk4Y9VjJA5IUKNZxB8AwfOS5g/MJ0L8Jx415beZG4jF3pLT6BM1na6oN/C
-IiBWwOM2/t9pq2zKr2hIoAHfCDTC8NCqsiXjYKgXBPa0d2p3bUthE9xjjUJ3Sny+
-XvtGK9sxzHt7N4srLubZfVf8h/a6WgvloZylhN6r1VuQI7bBbxZD+6mOWMOsXZCL
-YpyLWEWAuok6NgA8oJvT/te4GO3nz33tA/Z8QruvFtmhpn3XgDkasCWZrNmuqpo/
-p3pzRDRehSgxoq0dITatxfesWf/FDqWqs8YyBFeJ2X93gY1C/1YEmFMq+DhsZVNs
-AHIEckpIIy6lPhRchYJH+8T1W6+W4NqYZgtKLbwX0Fhdsmjy6u3FV5BNmnluI8Hu
-wXtVZCDofLtsnl8NXpCEr+x+ntDa0aKztXLyLYl3tzVLzGJvTzBIdyjVlEQy5a3Q
-Uv+mRAfcG/Ah5qt4yOue78NR6OGDr623WNpyXxj52FMww+uA/j+QpqXQhD/e0SH2
-OB/FAkl5qFz1vSdlgOS77axCNU/SWf8+UxygVE8X1gwI8R5RDCA=
------END CERTIFICATE-----
