--- community_images/ghost/bitnami/certs/server.key
@@ -1,52 +1,0 @@
------BEGIN PRIVATE KEY-----
-MIIJQQIBADANBgkqhkiG9w0BAQEFAASCCSswggknAgEAAoICAQDBmEHK9nQGeRZT
-5uVC9nvpbc9zui6+gfcc4LysCkPJUIYWrcbUGRbE2TtyEKz6xGwqFfEI9QfbCuAw
-nDNZ5PamT/GfP4WntUWCR5Y4v8VEwPPb09tqzuQa/MkmzpG84NjK5CJL31a2iO18
-51q3rcGJdil7xEm+hKVmBLXFKVqzWaXNirc0dbbQGaecOIU1B1iFW8rSsMaaX2LM
-EN3yfbxVzndAKWUwc4NdA+MTzHbva8lDNy+OQCHZz1z/Nd86IIyWH36/GmY1MN6U
-EKAsK36JtK3RMhkE6xxQv/hp4amjIc8pYYmxX0T8YVF6dFM8ct1cKbUv8X/dJHrI
-x+Xg/LoQLspkqGQE+VMjfV14Yafby9TewhIN9oekqGYvzmbkPodX3M3TEWyCUmiB
-iYY3QSax6NdpHk2mLOI1ezJfnby3rytgeDJxNXuQgwc92bKJtc4IUXqQKO1DWO1r
-f6H85GBa2zN6WZ8RIB3mXtuDQOXRFIDN7vmh2Nmo6AYo9s6zegycbnO10UhdmL1z
-ZcaNA3vCIKRshRgMjhuh6hyxv2aIf2eHn6OqCYKN6rqZuPB+7yuuaVJwrPLbAum6
-BXv3GyV0RNnPnTdqx8hBMq6NpAU/SuHp/7m5T1PrcnuQe+5C6HE+i1BSxn54l2GA
-C8u1shefSp+gJVyDUIYkPJ4iTUa+sQIDAQABAoICAAEQ/qPIkcELX7Ucm3yv8OTL
-Rw5DI0jHXH4bOs7Ty6XLPMHZOTQn4zk4XRx2vQ0wkcv1jPtYpi9tsuhxo9LF8Rnf
-bNxMtEbFN9nM3SX44HRNf8phM182PolrhQwirSjWMDuF4C58z3R38p7yg0xpt5wz
-O21EGofjISobsoigfrBoOQKV1gaC83vX+FVi9lDhdnpdlBNvB9kH13p6SU6xpJzw
-YzXGRJlsuXeEsvaD9M/CzzwAjNGCxuz1OHTrbgxa64tUJxWUZqW6XC+Cm+9CWhMW
-HLL/TKX7SY+kTxD+Pjn40LRQpjqE5NNjEdjT1TWLCG8eh7H46ZRlRAvhFgIUQk4c
-C/QLPe9cq1wjl5jdIYR/sgVWiDOztDBU4MFW/Dl1zHbCCQn7r8v3fO+PPuBxZHb2
-61ni7MXAimPwR4pmrD+jNIisAXTdUuqho/ZBtA1ltfBi3tO70AXYy6vOWiy9U8Ex
-L07YTl1qP1gFKdW8Yd6FvD5HGb0n+MnooF1CuNC8olJym544tF3YiYvEqAanM+L8
-Slz2WtK7LOXTfn6aCrYeOBPWKiQ9Xoy3DV/r87+5JbQem87wz8wSHCxa+stI7Tlw
-lRaPVuTabSsiAuWFSC3REk1ORdHpZd2L35sIMvpywrlP/rNugCusBDWgoB+fkFwa
-yC4FXUGh/6ZwVzOM2b/FAoIBAQD5aPb3ibG6gVSgSYvzCBlS9z5nCcrs4UYjrRrP
-IIL5qDvDxE8bbTwfSOswCRef3wTxMwqN5YgJGjRIS6ggP4Ye1Xs3QK/tKbhukvmn
-3AZC5dY7B2VTKhi6u30asQRH7huFW1U3NS3wvTcpCrrA5DUXNUTwIMyLFMUNYdj/
-cn8iTrEy3cuh2H4BUbbkZ8McAQWVFteX9ubffDG+AbfsD3LBQ514YGC//8rl4EqP
-NvF9pR/P4DWcvr0iInulVsxLxC77TVf7fvzmQhtHFVtzXptRk91zhbW5VP1Oxl5+
-5Vi0Qh2R8upeWD29m5Qjb7JMxx/L+VC0ILI+9Zncl3W3TMVnAoIBAQDGtcCFnDOd
-VSKTg6UAb8z1TlstQ6nuEiIxsSCHB6jgOjG6J/7WPeylXT+0qJoe3Q0jYo6XB+dQ
-WyqQKwVYJyrcF5QHabEgw1afBmwk2rJrk70260sAe8jinUtGdjl/BWFqw48R9F4k
-M9GZMrdTnEULpOCvaVDmRgnacsCol3JtrvgMBsoLkBqEtI+vnEO24AKf3vQmuNet
-d7mlJCByrMcTzsc71+tRde/6Juo7r9rQP3Qjx6W0M32tjA4LpsCjDVUGg9tRQ/5R
-jY1T0kJfqtNPBt4Lwmus2K0Vas1tx+yEMencR9TDM5THa2dqTxKCZrdql47hFNDj
-RQ+zQDEojnQnAoIBAAWf7mHqww1XuVbcEcHrXtbWKJweRmtNMU3Aq6gcLZf/yUlC
-9h3COubhuBwfOF8somnmIsjMcmLyKlevDR2dESNE4UB/vvf0RuqHegXlGmcL0+cE
-5a4YdvJPJCUj9NrCbfB1/Nf/5QdW3RWZ3lhXTTM2ylU4CCTi1QbWPBSWunNFnw1s
-nxkJTt4sv5krFeVdyIhyOW597PsBTAXi+Ajk2EuvgK+kTMpHtIW2pzfPn2acALmp
-bswYKR/J80vpRzY2OHEGb4OGe6GdwAu8WZGigpI2aXofe9ewKP5LzvnBplx6WTxf
-ZBW4m7vY9GBQNo8wtnDdMIq/agWteEjXSZpWquUCggEAHsmGPK1I6fWI1+YrqHep
-/8/WHJtBNtukZKKW8ZTkO2mIztjniPYrNivOdQUya9m/8V2RziCxXCfj3lZsC451
-+4fDrapYgUUCpLzwAH3J6xUG2f12r2p+Y24PfalPmziQMWR/RD7iXyWQMfZDpTav
-E/DFxyfAGw169VeWNgjQ9+GGGQbsQwJeuXToVkMV0r2+alq25VD6SawlHUsNo/bV
-GWKOR91aXlKxjcZ9wCWs91r9eZWzPMpFjKAIhRGfAyRLqii/0EWCn8wYu4X2NrBW
-8QWVXakP2cXwe0MC9SFmqcy5+SE/+8aAJK2lZjk+GuojPOaigokhHDVVC2bEtw3r
-PQKCAQAOqNREccSB7u97Uqsv4ZKgvL6mCc67Fv6fhxb+4M11/ofjDLcHg7fzApwE
-mbbn+xdGvZUj9RllRljWCkfT5JKspAVyISckzk6rSZipNoiSyxmZqGXXkVhWdFeI
-DSccE8ON50M1zcsm5gRpQLsxhiXZEiT1Q741wP/TrtYI7WsLaaZgQkxnOllL0bzG
-IfN3qZgdM5LUICi4E7X+9a6htNijDlsT2ZfwM0hsWsC42RAZKu8R72u/KEJgg0sX
-UlRrMMPlpoghw+W9XAeoTwfCwqx05r4/wUgvZnxVxQAu52/AEIh9nF4zAhxQptxd
-SkPDx8P9ddfRQScW3lI+n4xCpCQz
------END PRIVATE KEY-----
