--- community_images/ghost/bitnami/tls_certs.yml
@@ -1,24 +1,0 @@
-apiVersion: cert-manager.io/v1
-kind: Certificate
-metadata:
-  name: localhost-cert
-spec:
-  commonName: localhost
-  duration: 2160h
-  isCA: false
-  issuerRef:
-    group: cert-manager.io
-    kind: Issuer
-    name: ci-ca-issuer
-  privateKey:
-    algorithm: RSA
-    encoding: PKCS1
-    size: 2048
-  renewBefore: 360h
-  secretName: localhost-server-tls
-  subject:
-    organizations:
-    - rapidfort
-  usages:
-  - server auth
-  - client auth
